+++ R/type_lines.R
@@ -109,6 +109,8 @@ data_lines = function(dodge = 0, fixed.dodge = FALSE, xlevels = NULL, xord = NUL
       xlabs = seq_along(xlvls)
       names(xlabs) = xlvls
       datapoints[["x"]] = as.integer(datapoints[["x"]])
+      # More generous padding if categorical x-axis; better plot aesthetic
+      settings[["type_hints"]][["pads_cat_axis"]] = TRUE
     } else {
       xlabs = NULL
     }
@@ -117,6 +119,8 @@ data_lines = function(dodge = 0, fixed.dodge = FALSE, xlevels = NULL, xord = NUL
       ylabs = seq_along(ylvls)
       names(ylabs) = ylvls
       datapoints[["y"]] = as.integer(datapoints[["y"]])
+      # More generous padding if categorical y-axis; better plot aesthetic
+      settings[["type_hints"]][["pads_cat_axis"]] = TRUE
     } else {
       ylabs = NULL
     }